Uppdateringar Admin vs Customer: Admin Login: Ändrat alla referenser från User till Admin samt ändrat userList till adminList. Även ändrat så att den ärver av interface IAdminLogin istället. Customer Login: Ändrat så att klassen ärver från ICustomerLogin ist för ILogin.
Delat upp IMeny i ICustomerMenu och IAdminMenu
