+++ RegisterNewCustomer.cs
@@ -8,7 +8,11 @@ namespace DotNetDynamosV2
 {
     internal class RegisterNewCustomer
     {
-        public static void RegisterCustomer(LoginSystem loginSystem)
+        /// <summary>
+        /// Ändrat parameter att ta emot Admin loggedinAdmin
+        /// </summary>
+        /// <param name="loggedInAdmin"></param>
+        public static void RegisterCustomer(Admin loggedInAdmin)
         {
             // Skapa en temporär användare för att lägga till i användarlistan
             Customer newUser = new Customer();
@@ -17,7 +21,7 @@ namespace DotNetDynamosV2
             Console.Write("Enter your username: ");
             string enteredUsername = Console.ReadLine();
             // Kontrollera om användarnamnet redan finns
-            if (DataManager.userList.ContainsKey(enteredUsername))
+            if (DataManager.customerList.ContainsKey(enteredUsername))
             {
                 Console.WriteLine("Username already exists. Please choose a different username.");
                 return; // Avsluta metoden om användarnamnet redan finns
@@ -63,9 +67,9 @@ namespace DotNetDynamosV2
             // Antag att nextAdID är deklarerat någonstans som en statisk variabel i RegisterNewCustomer-klassen
             int nextAdID = 1;
             newUser.IDNumber = nextAdID++;
-           
+
             // Lägg till den nya användaren i userList (dictionary)
-            DataManager.userList.Add(newUser.UserName, newUser);
+            DataManager.customerList.Add(newUser.UserName, newUser);
 
             Console.Clear();
 
@@ -77,6 +81,76 @@ namespace DotNetDynamosV2
             {
                 Console.WriteLine($"Successfully created Customer {newUser.UserName}");
             }
+
+            //public static void RegisterCustomer(LoginSystem loginSystem)
+            //{
+            //    // Skapa en temporär användare för att lägga till i användarlistan
+            //    Customer newUser = new Customer();
+            //    Console.WriteLine("Welcome to User Registration!");
+            //    // Få användarnamn från användaren
+            //    Console.Write("Enter your username: ");
+            //    string enteredUsername = Console.ReadLine();
+            //    // Kontrollera om användarnamnet redan finns
+            //    if (DataManager.userList.ContainsKey(enteredUsername))
+            //    {
+            //        Console.WriteLine("Username already exists. Please choose a different username.");
+            //        return; // Avsluta metoden om användarnamnet redan finns
+            //    }
+            //    newUser.UserName = enteredUsername;
+            //    // Få förnamn från användaren
+            //    Console.Write("Enter your first name: ");
+            //    newUser.FirstName = Console.ReadLine();
+            //    // Få efternamn från användaren
+            //    Console.Write("Enter your last name: ");
+            //    newUser.LastName = Console.ReadLine();
+            //    // Få lösenord från användaren
+            //    Console.Write("Password must contain:\n6-12 characters\nAt least one capital letter\nAt least one digit\nAt least one symbol\nEnter password: ");
+            //    newUser.PassWord = Console.ReadLine();
+            //    // Låt användaren välja roll
+            //    Console.Write("Choose user role:\n");
+            //    Console.Write("1. Admin\n");
+            //    Console.Write("2. Customer\n");
+            //    newUser.Accounts = new List<Account>();
+            //    newUser.TransactionHistory = new List<Transaction>();
+            //    string roleChoice = Console.ReadLine();
+
+            //    if (int.TryParse(roleChoice, out int roleNumber))
+            //    {
+            //        switch (roleNumber)
+            //        {
+            //            case 1:
+            //                newUser.UserRole = "Admin";
+            //                break;
+            //            case 2:
+            //                newUser.UserRole = "Customer";
+            //                break;
+            //            default:
+            //                Console.WriteLine("Invalid choice. Please enter '1' for Admin or '2' for Customer.");
+            //                break;
+            //        }
+            //    }
+            //    else
+            //    {
+            //        Console.WriteLine("Invalid input. Please enter a valid number.");
+            //    }
+
+            //    // Antag att nextAdID är deklarerat någonstans som en statisk variabel i RegisterNewCustomer-klassen
+            //    int nextAdID = 1;
+            //    newUser.IDNumber = nextAdID++;
+
+            //    // Lägg till den nya användaren i userList (dictionary)
+            //    DataManager.userList.Add(newUser.UserName, newUser);
+
+            //    Console.Clear();
+
+            //    if (newUser is Admin)
+            //    {
+            //        Console.WriteLine($"Successfully created Admin {newUser.UserName}");
+            //    }
+            //    else if (newUser is Customer)
+            //    {
+            //        Console.WriteLine($"Successfully created Customer {newUser.UserName}");
+            //    }
 
 
             // Visa användarinformation
